@@ -1,6 +1,6 @@
 module github.com/petercb/k3os-bin
 
-go 1.16
+go 1.21.9
 
 require (
 	github.com/ghodss/yaml v1.0.0
@@ -13,6 +13,15 @@ require (
 	github.com/urfave/cli v1.22.9
 	golang.org/x/sys v0.0.0-20220829200755-d48e67d00261
 	pault.ag/go/modprobe v0.1.2
+)
+
+require (
+	github.com/cpuguy83/go-md2man/v2 v2.0.0 // indirect
+	github.com/mattn/go-shellwords v1.0.5 // indirect
+	github.com/russross/blackfriday/v2 v2.0.1 // indirect
+	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
+	gopkg.in/yaml.v2 v2.4.0 // indirect
+	pault.ag/go/topsort v0.0.0-20160530003732-f98d2ad46e1a // indirect
 )
 
 require (
