@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"bytes"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"sort"
@@ -103,6 +102,7 @@ func ApplyK3S(cfg *config.CloudConfig, restart, install bool) error {
 		return nil
 	}
 
+	//nolint:gocritic
 	if k3sExists {
 		vars = append(vars, "INSTALL_K3S_SKIP_DOWNLOAD=true")
 		vars = append(vars, "INSTALL_K3S_BIN_DIR=/sbin")
@@ -122,14 +122,14 @@ func ApplyK3S(cfg *config.CloudConfig, restart, install bool) error {
 			args = append(args, "server")
 		}
 	} else {
-		vars = append(vars, fmt.Sprintf("K3S_URL=%s", cfg.K3OS.ServerURL))
+		vars = append(vars, "K3S_URL="+cfg.K3OS.ServerURL)
 		if len(args) == 0 {
 			args = append(args, "agent")
 		}
 	}
 
 	if cfg.K3OS.Token != "" {
-		vars = append(vars, fmt.Sprintf("K3S_TOKEN=%s", cfg.K3OS.Token))
+		vars = append(vars, "K3S_TOKEN="+cfg.K3OS.Token)
 	}
 
 	var labels []string
@@ -137,9 +137,9 @@ func ApplyK3S(cfg *config.CloudConfig, restart, install bool) error {
 		labels = append(labels, fmt.Sprintf("%s=%s", k, v))
 	}
 	if mode != "" {
-		labels = append(labels, fmt.Sprintf("k3os.io/mode=%s", mode))
+		labels = append(labels, "k3os.io/mode="+mode)
 	}
-	labels = append(labels, fmt.Sprintf("k3os.io/version=%s", version.Version))
+	labels = append(labels, "k3os.io/version="+version.Version)
 	sort.Strings(labels)
 
 	for _, l := range labels {
@@ -160,7 +160,7 @@ func ApplyK3S(cfg *config.CloudConfig, restart, install bool) error {
 	return cmd.Run()
 }
 
-func ApplyInstall(cfg *config.CloudConfig) error {
+func ApplyInstall(_ *config.CloudConfig) error {
 	mode, err := mode.Get()
 	if err != nil {
 		return err
@@ -197,9 +197,9 @@ func ApplyDNS(cfg *config.CloudConfig) error {
 		buf.WriteString("\n")
 	}
 
-	err := ioutil.WriteFile("/etc/connman/main.conf", buf.Bytes(), 0644)
+	err := os.WriteFile("/etc/connman/main.conf", buf.Bytes(), 0644)
 	if err != nil {
-		return fmt.Errorf("failed to write /etc/connman/main.conf: %v", err)
+		return fmt.Errorf("failed to write /etc/connman/main.conf: %w", err)
 	}
 
 	return nil
@@ -218,10 +218,10 @@ func ApplyWifi(cfg *config.CloudConfig) error {
 
 	if buf.Len() > 0 {
 		if err := os.MkdirAll("/var/lib/connman", 0755); err != nil {
-			return fmt.Errorf("failed to mkdir /var/lib/connman: %v", err)
+			return fmt.Errorf("failed to mkdir /var/lib/connman: %w", err)
 		}
-		if err := ioutil.WriteFile("/var/lib/connman/settings", buf.Bytes(), 0644); err != nil {
-			return fmt.Errorf("failed to write to /var/lib/connman/settings: %v", err)
+		if err := os.WriteFile("/var/lib/connman/settings", buf.Bytes(), 0644); err != nil {
+			return fmt.Errorf("failed to write to /var/lib/connman/settings: %w", err)
 		}
 	}
 
@@ -246,7 +246,7 @@ func ApplyWifi(cfg *config.CloudConfig) error {
 	}
 
 	if buf.Len() > 0 {
-		return ioutil.WriteFile("/var/lib/connman/cloud-config.config", buf.Bytes(), 0644)
+		return os.WriteFile("/var/lib/connman/cloud-config.config", buf.Bytes(), 0644)
 	}
 
 	return nil
@@ -264,8 +264,8 @@ func ApplyDataSource(cfg *config.CloudConfig) error {
 	buf.WriteString(args)
 	buf.WriteString("\"\n")
 
-	if err := ioutil.WriteFile("/etc/conf.d/cloud-config", buf.Bytes(), 0644); err != nil {
-		return fmt.Errorf("failed to write to /etc/conf.d/cloud-config: %v", err)
+	if err := os.WriteFile("/etc/conf.d/cloud-config", buf.Bytes(), 0644); err != nil {
+		return fmt.Errorf("failed to write to /etc/conf.d/cloud-config: %w", err)
 	}
 
 	return nil
@@ -276,7 +276,7 @@ func ApplyEnvironment(cfg *config.CloudConfig) error {
 		return nil
 	}
 	env := make(map[string]string, len(cfg.K3OS.Environment))
-	if buf, err := ioutil.ReadFile("/etc/environment"); err == nil {
+	if buf, err := os.ReadFile("/etc/environment"); err == nil {
 		scanner := bufio.NewScanner(bytes.NewReader(buf))
 		for scanner.Scan() {
 			line := scanner.Text()
@@ -309,8 +309,8 @@ func ApplyEnvironment(cfg *config.CloudConfig) error {
 		buf.WriteString(strconv.Quote(val))
 		buf.WriteString("\n")
 	}
-	if err := ioutil.WriteFile("/etc/environment", buf.Bytes(), 0644); err != nil {
-		return fmt.Errorf("failed to write to /etc/environment: %v", err)
+	if err := os.WriteFile("/etc/environment", buf.Bytes(), 0644); err != nil {
+		return fmt.Errorf("failed to write to /etc/environment: %w", err)
 	}
 
 	return nil
