@@ -2,7 +2,6 @@ package cliinstall
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"os/exec"
 
@@ -79,7 +78,7 @@ func runInstall(cfg config.CloudConfig) error {
 	}
 
 	if cfg.K3OS.Install.ConfigURL == "" {
-		tempFile, err = ioutil.TempFile("/tmp", "k3os.XXXXXXXX")
+		tempFile, err = os.CreateTemp("/tmp", "k3os.XXXXXXXX")
 		if err != nil {
 			return err
 		}
