@@ -3,7 +3,6 @@ package config
 import (
 	"bytes"
 	"encoding/base64"
-	"io/ioutil"
 	"os"
 	"strings"
 
@@ -21,12 +20,12 @@ func readCloudConfig() (map[string]interface{}, error) {
 	var keys []string
 	result := map[string]interface{}{}
 
-	hostname, err := ioutil.ReadFile(hostname)
+	hostname, err := os.ReadFile(hostname)
 	if err == nil {
 		result["hostname"] = strings.TrimSpace(string(hostname))
 	}
 
-	keyData, err := ioutil.ReadFile(ssh)
+	keyData, err := os.ReadFile(ssh)
 	if err != nil {
 		// ignore error
 		return result, nil
@@ -49,7 +48,7 @@ func readCloudConfig() (map[string]interface{}, error) {
 func readUserData() (map[string]interface{}, error) {
 	result := map[string]interface{}{}
 
-	data, err := ioutil.ReadFile(userdata)
+	data, err := os.ReadFile(userdata)
 	if os.IsNotExist(err) {
 		return nil, nil
 	} else if err != nil {
