@@ -1,7 +1,6 @@
 package mode
 
 import (
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strings"
@@ -10,7 +9,7 @@ import (
 )
 
 func Get(prefix ...string) (string, error) {
-	bytes, err := ioutil.ReadFile(filepath.Join(filepath.Join(prefix...), system.StatePath("mode")))
+	bytes, err := os.ReadFile(filepath.Join(filepath.Join(prefix...), system.StatePath("mode")))
 	if os.IsNotExist(err) {
 		return "", nil
 	} else if err != nil {
