@@ -2,7 +2,7 @@ package ssh
 
 import (
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/url"
 	"os"
@@ -22,7 +22,7 @@ const (
 )
 
 func SetAuthorizedKeys(cfg *config.CloudConfig, withNet bool) error {
-	bytes, err := ioutil.ReadFile("/etc/passwd")
+	bytes, err := os.ReadFile("/etc/passwd")
 	if err != nil {
 		return err
 	}
@@ -73,7 +73,7 @@ func getKey(key string, withNet bool) (string, error) {
 	var resp *http.Response
 	for i := 0; i < 10; time.Sleep(time.Second) {
 		// network interface(s) can be up before DNS is ready, so let's try up to 10 times
-		resp, err = http.Get(key)
+		resp, err = http.Get(key) //nolint:bodyclose
 		if err == nil || strings.Contains(err.Error(), "unsupported protocol scheme") {
 			break
 		}
@@ -88,7 +88,7 @@ func getKey(key string, withNet bool) (string, error) {
 	if resp.StatusCode/100 > 2 {
 		return "", fmt.Errorf("%s %s", resp.Proto, resp.Status)
 	}
-	bytes, err := ioutil.ReadAll(resp.Body)
+	bytes, err := io.ReadAll(resp.Body)
 	return string(bytes), err
 }
 
@@ -117,7 +117,7 @@ func authorizeSSHKey(key, file string, uid, gid int, withNet bool) error {
 	} else if err != nil {
 		return err
 	}
-	bytes, err := ioutil.ReadFile(file)
+	bytes, err := os.ReadFile(file)
 	if err != nil {
 		return err
 	}
