@@ -2,7 +2,7 @@ package util
 
 import (
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"os"
 	"os/exec"
@@ -11,7 +11,7 @@ import (
 
 func WriteFileAtomic(filename string, data []byte, perm os.FileMode) error {
 	dir, file := path.Split(filename)
-	tempFile, err := ioutil.TempFile(dir, fmt.Sprintf(".%s", file))
+	tempFile, err := os.CreateTemp(dir, "."+file)
 	if err != nil {
 		return err
 	}
@@ -34,7 +34,7 @@ func HTTPDownloadToFile(url, dest string) error {
 		return err
 	}
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
+	body, err := io.ReadAll(res.Body)
 	if err != nil {
 		return err
 	}
@@ -50,7 +50,7 @@ func HTTPLoadBytes(url string) ([]byte, error) {
 			return nil, fmt.Errorf("non-200 http response: %d", resp.StatusCode)
 		}
 
-		bytes, err := ioutil.ReadAll(resp.Body)
+		bytes, err := io.ReadAll(resp.Body)
 		if err != nil {
 			return nil, err
 		}
